#27: set signup/in request
