+++ app/assets/stylesheets/application.css
@@ -21,10 +21,12 @@
  }
 
  #title {
+     float: left;
      height: 20%;
      width: 100%;
  }
  .workspace {
+     clear: left;
      display: flex;
      flex-wrap: wrap;
      flex-basis: 66%;
@@ -259,15 +261,6 @@
 
 /* style of signUp, signIn windows */
 
-.signup-window {
-  background-color: #FFFFFF;
-  width: 400px;
-  height: 400px;
-  margin: 7em auto;
-  box-shadow: 0px 11px 35px 2px rgba(0, 0, 0, 0.14);
-  visibility: hidden;
-}
-
 .signin-window {
   display: none;
   background-color: #FFFFFF;
@@ -278,6 +271,18 @@
   top: 50%;
   transform: translate(-50%, -50%);
 }
+
+.signup-window {
+  display: none;
+  background-color: #FFFFFF;
+  height: 400px;
+  width: 400px;
+  position: absolute;
+  left: 50%;
+  top: 50%;
+  transform: translate(-50%, -50%);
+}
+
 
 .show {
   display: block;
@@ -302,6 +307,6 @@
   width: 30px;
 }
 
-/* .auth-links {
-  display: none;
-} */
+.auth-links {
+  float: right;
+}
