+++ app/views/welcome/index.html.erb
@@ -1,7 +1,9 @@
 <div class="auth-links">
-  <a id="signup-link" class="sign-links" href="#">Signup</a>
-  <a id="signin-link" class="sign-links" href="#">Signin</a>
-  <a id="signout-link" class="sign-links" href="#">Signout</a>
+  <right>
+  <a id="signup-link" class="sign-links" href="#">Sign Up</a>
+  <a id="signin-link" class="sign-links" href="#">Sign In</a>
+  <a id="signout-link" class="sign-links" href="#">Sign Out</a>
+  </right>
 </div>
 
 <div id="title">
@@ -25,25 +27,28 @@
 </div>
 
 <!-- signup window -->
-<!-- <div class="signup-window">
+<div class="signup-window">
   <div class="close-window-icon"></div>
-  <center><h2>Sign Up:</h2></center>
-  <form action="#">
-    <label for="email">email:</label><br>
-    <input type="text" name="email" value=""><br>
-    <label for="password">password:</label><br>
-    <input type="password" name="password" value=""><br><br>
-    <label for="password_confirmation">confirm password:</label><br>
-    <input type="password" name="password_confirmation" value=""><br><br>
-    <input type="submit" id="signup-submit" value="Submit">
-</form>
-</div> -->
+  <center>
+    <h2>Sign Up:</h2>
+    <form action="#">
+      <label for="email">email:</label><br>
+      <input type="text" name="email" class="input-email" value=""><br>
+      <label for="password">password:</label><br>
+      <input type="password" name="password" class="input-password" value=""><br><br>
+      <label for="password_confirmation">confirm password:</label><br>
+      <input type="password" name="password_confirmation" class="input-password-confirmation" value=""><br><br>
+      <input type="submit" id="signup-submit" value="Submit">
+    </form>
+  </center>
+</div>
 
  <!-- signin window -->
 <div class="signin-window">
   <div class="close-window-icon"></div>
   <center><h2>Sign In:</h2></center>
   <div>
+    <center>
     <form action="#" >
       <label for="email">email:</label><br>
       <input type="text" name="email" value="" class="input-email"><br>
@@ -51,5 +56,6 @@
       <input type="password" name="password" value="" class="input-password"><br><br>
       <input type="submit" id="signin-submit" value="Submit">
   </form>
+  </center>
   </div>
 </div>
